atualizacao29_04_23

+++ database/migrations/20221117035111-create-paper.js
@@ -35,7 +35,7 @@ module.exports = {
         allowNull: true,
       },
       price: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DOUBLE,
         allowNull: true,
       },
       efficiency: {
@@ -43,7 +43,7 @@ module.exports = {
         allowNull: true,
       },
       unitprice: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DOUBLE,
         allowNull: true,
       },
       provider_id: {
