+++ database/migrations/20230108223715-create-budgetmaterial.js
@@ -27,7 +27,7 @@ module.exports = {
         allowNull: true
       },
       amount: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DOUBLE,
         allowNull: true,
       },
       created_at: {
